requirements.txt
chainer==7.7.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
filelock==3.0.12
imageio==2.9.0
kiwisolver==1.3.1
kornia==0.5.2
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.52
pandas==1.2.4
pandas-stubs==1.1.0.7
Pillow==8.2.0
protobuf==3.17.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
